Apply patch for cross-compiling ARM64 lkl from linux kernel 4.4
The MPTCP supported kernel version is 4.4. LKL has been back-ported
to be used against this one.Cross-compiling LKL here with the
Linaro aarch64-linux-android-5.4-linaro toolchain.

+++ tools/lkl/lib/hijack/hijack.c
@@ -193,7 +193,7 @@ int socket(int domain, int type, int protocol)
 }
 
 HOST_CALL(ioctl);
-int ioctl(int fd, unsigned long req, ...)
+int ioctl(int fd, int req, ...)
 {
 	va_list vl;
 	long arg;
